Gjør apiInternal til ett source-set bare slik at de kan dele internals (intellij har ikke støtte for "kotlin-friends" enda). Splitter inn i api-internal og external (bare src).
