+++ brevbaker/markup/build.gradle.kts
@@ -26,6 +26,28 @@ dependencies {
     testImplementation(libs.bundles.junit)
 }
 
+// The `apiInternal` source set is an internal-API layer that may use markup's `internal` seams
+// (builders, IdGenerator, markupJson, internal constructors). It is kept as a dedicated source set
+// — rather than folded into `main` — so those seams do NOT become part of markup's published public
+// API (the binary-compatibility validator and the published jar only cover `main`).
+//
+// It gains `internal` access to `main` through Kotlin compilation association (`associateWith`)
+// instead of a raw cross-module `friendPaths` jar. Both grant `internal` access at compile time,
+// but only associateWith is exported into the IntelliJ Gradle model, so the IDE's code analysis
+// recognizes the friend relationship and stops flagging the internal usages as errors (a plain
+// cross-module friendPaths entry is not understood by IntelliJ — see KT-35347).
+sourceSets.create("apiInternal")
+
+kotlin {
+    val mainCompilation = target.compilations.getByName("main")
+    val apiInternalCompilation = target.compilations.getByName("apiInternal")
+    apiInternalCompilation.associateWith(mainCompilation)
+
+    // Tests exercise the apiInternal DSL, so the test compilation needs it (and transitively main)
+    // on its classpath, plus friend access.
+    target.compilations.getByName("test").associateWith(apiInternalCompilation)
+}
+
 tasks.test {
     useJUnitPlatform()
 }
@@ -67,6 +89,9 @@ tasks {
         targetCompatibility = markupJavaTarget
     }
     compileTestJava {
+        targetCompatibility = markupJavaTarget
+    }
+    named<JavaCompile>("compileApiInternalJava") {
         targetCompatibility = markupJavaTarget
     }
 }
