+++ brevbaker/markup/src/main/kotlin/no/nav/brev/brevbaker/markup/dsl/ContentBuilder.kt
@@ -1,6 +1,5 @@
 package no.nav.brev.brevbaker.markup.dsl
 
-import no.nav.brev.brevbaker.markup.ElementTags
 import no.nav.brev.brevbaker.markup.outline.Text
 import no.nav.brev.brevbaker.markup.outline.Text.FontType
 
@@ -11,8 +10,7 @@ import no.nav.brev.brevbaker.markup.outline.Text.FontType
  * - [text] for vanlig tekst
  * - [newLine] for linjeskift
  *
- * Bruk [ExtendedContentBuilder] (via `*Extended`-inngangene) når du også trenger [ExtendedContentBuilder.variable]
- * og støtte for metadata som tags.
+ * En utvidet variant med `variable` og metadata (tags) finnes i api-internal-modulen.
  */
 @MarkupDsl
 abstract class AbstractContentBuilder internal constructor(private val ids: IdGenerator) {
@@ -54,56 +52,6 @@ abstract class AbstractContentBuilder internal constructor(private val ids: IdGe
 @MarkupDsl
 class ContentBuilder internal constructor(ids: IdGenerator) : AbstractContentBuilder(ids)
 
-/** Tekstinnhold som også tillater [variable] og tags. */
-@MarkupDsl
-class ExtendedContentBuilder internal constructor(ids: IdGenerator) : AbstractContentBuilder(ids) {
-    /**
-     * Legg til en variabel (datafelt som fylles ut ved rendring), valgfritt med [fontType] og [tags].
-     *
-     * ```
-     * text("Du får ")
-     * variable("1000 kr", tags = setOf(ElementTags.REDIGERBAR_DATA))
-     * ```
-     */
-    fun variable(text: String, fontType: FontType = FontType.PLAIN, tags: Set<ElementTags> = emptySet()) {
-        texts.add(Text.Variable(nextId(), text, fontType, tags))
-    }
-}
-
-/**
- * Begrenset tekst-scope for overskrifter/ledetekster (titler, kolonneoverskrifter, form-choice-ledetekst).
- * Tillater kun ren tekst: [text] og [variable] uten [FontType] (alltid [FontType.PLAIN]) og uten linjeskift.
- */
-@MarkupDsl
-class PlainExtendedTextBuilder internal constructor(private val ids: IdGenerator) {
-    private val texts: MutableList<Text> = mutableListOf()
-
-    /**
-     * Vanlig tekst.
-     *
-     * ```
-     * text("Vedtak")
-     * ```
-     */
-    fun text(text: String) {
-        texts.add(Text.Literal(ids.next(), text, FontType.PLAIN))
-    }
-
-    /**
-     * Legg til en variabel.
-     *
-     * ```
-     * text("Vedtak for ")
-     * variable("navn")
-     * ```
-     */
-    fun variable(text: String) {
-        texts.add(Text.Variable(ids.next(), text, FontType.PLAIN))
-    }
-
-    internal fun build(): List<Text> = texts.toList()
-}
-
 /** Begrenset tekst-scope uten `variable` for plain-only overskrifter i [letterMarkup]. */
 @MarkupDsl
 class PlainTextBuilder internal constructor(private val ids: IdGenerator) {
@@ -134,7 +82,3 @@ internal fun IdGenerator.plainText(text: String): List<Text> = listOf(Text.Liter
 /** Ren tekst fra DSL-blokk (kun [Text.Literal], [FontType.PLAIN], ingen linjeskift). */
 internal fun IdGenerator.plainText(build: PlainTextBuilder.() -> Unit): List<Text> =
     PlainTextBuilder(this).apply(build).build()
-
-/** Ren tekst med `variable` (kun [Text.Literal]/[Text.Variable], [FontType.PLAIN], ingen linjeskift). */
-internal fun IdGenerator.plainExtendedText(build: PlainExtendedTextBuilder.() -> Unit): List<Text> =
-    PlainExtendedTextBuilder(this).apply(build).build()
