+++ brevbaker/markup/src/test/kotlin/no/nav/brev/brevbaker/markup/LetterMarkupDslTest.kt
@@ -1,7 +1,5 @@
 package no.nav.brev.brevbaker.markup
 
-import no.nav.brev.brevbaker.markup.ElementTags.FRITEKST
-import no.nav.brev.brevbaker.markup.dsl.column
 import no.nav.brev.brevbaker.markup.dsl.*
 import no.nav.brev.brevbaker.markup.outline.Block
 import no.nav.brev.brevbaker.markup.outline.Block.FormText.Size
@@ -14,114 +12,6 @@ import org.junit.jupiter.api.Test
 import java.time.LocalDate
 
 class LetterMarkupDslTest {
-
-    private fun fullLetter(): LetterMarkup = letterMarkupExtended {
-        title1 { text("Vedtak om uføretrygd") }
-        saksinformasjon(
-            gjelderNavn = "Ola Nordmann",
-            gjelderFoedselsnummer = "12345678901",
-            saksnummer = "9876543",
-            dokumentDato = LocalDate.of(2026, 7, 9),
-            annenMottakerNavn = "Kari Nordmann",
-        )
-        outline {
-            title2 { text("Innledning") }
-            paragraph {
-                text("Du får ")
-                newLine()
-                variable("uføretrygd")
-                text(".")
-            }
-            itemList {
-                item { text("Punkt 1") }
-                item { text("Punkt 2") }
-                item("")
-            }
-            numberedList {
-                item {
-                    text("Steg 1")
-                }
-            }
-            table {
-                header {
-                    column(ColumnAlignment.LEFT) {
-                        text("Kolonne ")
-                        variable("A")
-                    }
-                    column(ColumnAlignment.RIGHT, span = 2) { text("Kolonne B") }
-                }
-                row {
-                    cell { text("A1") }
-                    cell { text("B1") }
-                }
-            }
-            formText(Size.LONG) { text("Skriv her") }
-            formChoice {
-                prompt { text("Velg") }
-                choice { text("Ja") }
-                choice { text("Nei") }
-            }
-        }
-        signatur(
-            hilsenTekst = "Med vennlig hilsen",
-            navAvsenderEnhet = "NAV Familie- og pensjonsytelser",
-            saksbehandlerNavn = "Sak S.Behandler",
-            attesterendeSaksbehandlerNavn = "Att Esterer",
-        )
-    }
-
-    @Test
-    fun `dsl builds a valid letter with all block and text types`() {
-        val letter = fullLetter()
-
-        assertEquals(2, letter.version)
-        assertEquals("Ola Nordmann", letter.saksinformasjon.gjelderNavn)
-        assertEquals("12345678901", letter.saksinformasjon.gjelderFoedselsnummer.value)
-        assertEquals("9876543", letter.saksinformasjon.saksnummer.saksnummer)
-
-        val types = letter.blocks.map { it.type }
-        assertTrue(
-            types.containsAll(
-                listOf(
-                    Block.Type.TITLE2,
-                    Block.Type.PARAGRAPH,
-                    Block.Type.ITEM_LIST,
-                    Block.Type.NUMBERED_LIST,
-                    Block.Type.TABLE,
-                    Block.Type.FORM_TEXT,
-                    Block.Type.FORM_CHOICE,
-                )
-            )
-        )
-
-        val paragraph = letter.blocks.filterIsInstance<Block.Paragraph>().single()
-        val textTypes = paragraph.content.map { it.type }
-        assertTrue(textTypes.containsAll(listOf(Text.Type.LITERAL, Text.Type.VARIABLE, Text.Type.NEW_LINE)))
-    }
-
-    @Test
-    fun `dsl assigns unique ids`() {
-        val letter = fullLetter()
-        val ids = mutableListOf<Int>()
-        letter.title1.forEach { ids.add(it.id) }
-        letter.blocks.forEach { block ->
-            ids.add(block.id)
-            when (block) {
-                is Block.Title2 -> block.content.forEach { ids.add(it.id) }
-                is Block.Paragraph -> block.content.forEach { ids.add(it.id) }
-                else -> {}
-            }
-        }
-        assertEquals(ids.size, ids.toSet().size, "ids should be unique")
-    }
-
-    @Test
-    fun `letter round-trips through json`() {
-        val letter = fullLetter()
-        val json = letter.toJson()
-        val decoded = decodeLetterMarkup(json)
-        assertEquals(letter, decoded)
-    }
 
     @Test
     fun `base letterMarkup builds without variables`() {
@@ -190,105 +80,25 @@ class LetterMarkupDslTest {
     }
 
     @Test
-    fun `extended DSL supports shorthand string methods with font type on content`() {
-        val letter = letterMarkupExtended {
-            title1("Vedtak")
-            title1 {
-                text("Tittel fra builder ")
-                variable("x")
-            }
+    fun `base letter round-trips through json`() {
+        val letter = letterMarkup {
             saksinformasjon(
                 gjelderNavn = "Ola Nordmann",
                 gjelderFoedselsnummer = "12345678901",
                 saksnummer = "9876543",
                 dokumentDato = LocalDate.of(2026, 7, 9),
             )
+            title1("Tittel")
             outline {
-                title2("Innledning")
-                title2 {
-                    text("Innledning ")
-                    variable("x")
-                }
-                title3("Mellomtittel")
-                formText("TEXT", Size.SHORT)
-                title3 {
-                    text("Mellomtittel ")
-                    variable("x")
-                }
-                title4("Detaljer")
-                title4 {
-                    text("Detaljer ")
-                    variable("x")
-                }
-                paragraph("Ingress", fontType = FontType.BOLD)
-                paragraph {
-                    text("Du får ")
-                    variable("1000 Kr", tags = setOf(ElementTags.REDIGERBAR_DATA))
-                    newLine()
-                }
-                itemList {
-                    item("Punkt", fontType = FontType.BOLD)
-                    item {
-                        text("Du får ")
-                        variable("1000 Kr")
-                    }
-                }
-                table {
-                    header {
-                        column("Kolonne")
-                        column {
-                            text("Kolonne ")
-                            variable("2")
-                        }
-                    }
-                    row {
-                        cell("Celle", fontType = FontType.BOLD)
-                        cell {
-                            text("bla")
-                            variable("2")
-                        }
-                    }
-                }
-                formText("Ledetekst", Size.SHORT, fontType = FontType.BOLD)
-                formText(Size.LONG) {
-                    text("Skriv ")
-                    variable("her")
-                }
-                formChoice {
-                    prompt("Velg")
-                    prompt {
-                        text("Du får svar innen ")
-                        variable("x")
-                        text("uker.")
-                    }
-                    choice("Ja", fontType = FontType.BOLD)
-                    choice("Nei")
-                    choice {
-                        text("kanskje")
-                        variable("eller?", tags = setOf(FRITEKST))
-                    }
-                }
+                paragraph("Innhold")
             }
             signatur(
-                hilsenTekst = "Hilsen",
+                hilsenTekst = "Med vennlig hilsen",
                 navAvsenderEnhet = "NAV",
             )
         }
-
-        val titleTexts = letter.title1.filterIsInstance<Text.Literal>().map { it.text }
-        assertTrue(titleTexts.contains("Tittel fra builder "))
-
-        val title2 = letter.blocks.filterIsInstance<Block.Title2>().first()
-        assertEquals(FontType.PLAIN, title2.content.filterIsInstance<Text.Literal>().first().fontType)
-
-        val paragraph = letter.blocks.filterIsInstance<Block.Paragraph>().first()
-        assertEquals(FontType.BOLD, (paragraph.content.single() as Text.Literal).fontType)
-
-        val item = letter.blocks.filterIsInstance<Block.ItemList>().single().items.first()
-        assertEquals(FontType.BOLD, (item.content.single() as Text.Literal).fontType)
-
-        val cell = letter.blocks.filterIsInstance<Block.Table>().single().rows.single().cells.first()
-        assertEquals(FontType.BOLD, (cell.text.single() as Text.Literal).fontType)
+        val decoded = decodeLetterMarkup(letter.toJson())
+        assertEquals(letter, decoded)
     }
 
     @Test
@@ -300,55 +110,49 @@ class LetterMarkupDslTest {
     }
 
     @Test
-    fun `pdfTittelExtended supports variables and round-trips through json`() {
-        val tittel = pdfTittelExtended {
-            text("Vedtak for ")
-            variable("Ola Nordmann")
-        }
-
-        assertEquals(listOf(Text.Type.LITERAL, Text.Type.VARIABLE), tittel.title1.map { it.type })
-
-        val decoded = decodePDFTittel(tittel.toJson())
-        assertEquals(tittel, decoded)
-    }
-
-    @Test
-    fun `letterMarkupWithDataUsage includes brevtype and data usage`() {
-        val property = dataUsageProperty(typeName = "UngUfoerDto", propertyName = "totaltUfoerePerMnd")
-
-        val letter = letterMarkupWithDataUsage(
+    fun `letterPDFRequest builds and round-trips through json`() {
+        val request = letterPDFRequest(
+            language = LanguageCode.BOKMAL,
             brevtype = Brevtype.VEDTAKSBREV,
-            letterDataUsage = setOf(property),
         ) {
-            saksinformasjon(
-                gjelderNavn = "Ola Nordmann",
-                gjelderFoedselsnummer = "12345678901",
-                saksnummer = "9876543",
-                dokumentDato = LocalDate.of(2026, 7, 9),
-            )
-            title1("Vedtak")
-            outline {
-                paragraph("Innhold")
+            letter {
+                saksinformasjon(
+                    gjelderNavn = "Ola Nordmann",
+                    gjelderFoedselsnummer = "12345678901",
+                    saksnummer = "9876543",
+                    dokumentDato = LocalDate.of(2026, 7, 9),
+                )
+                title1("Vedtak")
+                outline {
+                    paragraph("Innhold")
+                }
+                signatur(
+                    hilsenTekst = "Med vennlig hilsen",
+                    navAvsenderEnhet = "NAV",
+                )
             }
-            signatur(
-                hilsenTekst = "Med vennlig hilsen",
-                navAvsenderEnhet = "NAV",
-            )
+            attachment(inkluderSaksinformasjon = true) {
+                title1("Vedlegg 1")
+                outline {
+                    paragraph("Vedleggsinnhold")
+                }
+            }
+            pdfVedlegg { text("Ekstra PDF-vedlegg") }
         }
 
-        assertEquals(Brevtype.VEDTAKSBREV, letter.brevtype)
-        assertEquals(setOf(property), letter.letterDataUsage)
-        assertEquals("Vedtak", (letter.markup.title1.single() as Text.Literal).text)
+        assertEquals(LanguageCode.BOKMAL, request.language)
+        assertEquals(Brevtype.VEDTAKSBREV, request.brevtype)
+        assertEquals("Vedtak", (request.letterMarkup.title1.single() as Text.Literal).text)
+        assertEquals(1, request.attachments.size)
+        assertEquals("Ekstra PDF-vedlegg", (request.pdfVedlegg.single().title1.single() as Text.Literal).text)
+
+        val decoded = decodeLetterPDFRequest(request.toJson())
+        assertEquals(request, decoded)
     }
 
     @Test
-    fun `letterMarkupWithDataUsageExtended supports variable content and round-trips through json`() {
-        val property = dataUsageProperty(typeName = "UngUfoerDto", propertyName = "belop")
-
-        val letter = letterMarkupWithDataUsageExtended(
-            brevtype = Brevtype.INFORMASJONSBREV,
-            letterDataUsage = setOf(property),
-        ) {
+    fun `letterPDFRequest accepts pre-built components`() {
+        val letter = letterMarkup {
             saksinformasjon(
                 gjelderNavn = "Ola Nordmann",
                 gjelderFoedselsnummer = "12345678901",
@@ -356,23 +160,21 @@ class LetterMarkupDslTest {
                 dokumentDato = LocalDate.of(2026, 7, 9),
             )
             title1("Orientering")
-            outline {
-                paragraph {
-                    text("Beløp: ")
-                    variable("1000 kr")
-                }
-            }
-            signatur(
-                hilsenTekst = "Med vennlig hilsen",
-                navAvsenderEnhet = "NAV",
-            )
+            outline { paragraph("Innhold") }
+            signatur(hilsenTekst = "Med vennlig hilsen", navAvsenderEnhet = "NAV")
+        }
+        val vedlegg = attachment { title1("Vedlegg"); outline { paragraph("X") } }
+        val tittel = pdfTittel { text("Tittel") }
+
+        val request = letterPDFRequest(language = LanguageCode.NYNORSK, brevtype = Brevtype.INFORMASJONSBREV) {
+            letter(letter)
+            attachment(vedlegg)
+            pdfVedlegg(tittel)
         }
 
-        val paragraph = letter.markup.blocks.single() as Block.Paragraph
-        assertEquals(listOf(Text.Type.LITERAL, Text.Type.VARIABLE), paragraph.content.map { it.type })
-
-        val decoded = decodeLetterMarkupWithDataUsage(letter.toJson())
-        assertEquals(letter, decoded)
+        assertEquals(letter, request.letterMarkup)
+        assertEquals(listOf(vedlegg), request.attachments)
+        assertEquals(listOf(tittel), request.pdfVedlegg)
     }
 
     @Test
@@ -388,23 +190,5 @@ class LetterMarkupDslTest {
         assertEquals("Vedlegg 1", (vedlegg.title1.single() as Text.Literal).text)
         val paragraph = vedlegg.blocks.single() as Block.Paragraph
         assertEquals("Innhold", (paragraph.content.single() as Text.Literal).text)
-    }
-
-    @Test
-    fun `attachmentExtended supports variable text`() {
-        val vedlegg = attachmentExtended(inkluderSaksinformasjon = false) {
-            title1("Vedlegg 2")
-            outline {
-                paragraph {
-                    text("Sats ")
-                    variable("2G")
-                }
-            }
-        }
-
-        assertEquals(false, vedlegg.inkluderSaksinformasjon)
-        val paragraph = vedlegg.blocks.single() as Block.Paragraph
-        val textTypes = paragraph.content.map { it.type }
-        assertTrue(textTypes.containsAll(listOf(Text.Type.LITERAL, Text.Type.VARIABLE)))
     }
 }
